Add stacked piece checking to Rook's move method.
Add a check for if the player attempts to move their rook onto a
tile occupied by another of that same player's pieces.

Also add tracking for which piece is occupying the tile the rook is
trying to move onto. This will be useful for when piece captures
are implemented for the Rook class.