Reorganize project into a single package.
Since the code in the ChessPiece package was already tightly coupled
to the code outside of the package, and reusing the ChessPiece package
else-where was probably not going to happen, moving everything to the
same package makes sense.

This will reduce the complexity of programming this project, since
everything that was inside the ChessPiece package is now outside in
the same package as the rest of the program.