Redid Castling With If Statement
Took out the first for loop in the king class that found the index of
the king and replaced it with if else statements. Added in more
parameters for the move method in king for the constants that store
indexs. Added in extra move method in abstract class which contains
extra parameters. Fixed the if statements for rook so calling the set
not moved method would actually work. Added in if else in board.java
to determine which move method to call.