Implementation of Checkmate
Created the checkmate method along with changing the check method in
the king class. Along wiht setCheckCaptured and setChecker methods for
necessary variables for checkmate. Set the check captured in each of the
piece classes whenever they capture. Fixed bug where the captured piece
counts did not initialize back at 0 when reset. Fixed bugs in the
notation with the move type.