Fix bug preventing King from capturing Bishops.
Fix a bug which was preventing the King from capturing Bishops, the root
cause of which was the Bishop being captured reporting that it could
move onto the square it was already on to the check method, which caused
the check method to report that the King was in check after performing
the move.

Also cleaned up an incorrect comment in King.java.