Implement Network Play with Bug Fix
Fixed a checkmate bug that was happening when a king in check could
capture a piece to get out of checkmate. Implemented the server and
client threads used for network play. Changed the mouse released method
in BoardPanel so that communication between server and cliet is possible
added in extra functionality with the forfeit and new game buttons in
the menu panel for network play.