Fix JUnit tests failing.
Fix JUnit tests failing, which were failing due to not finding the image
files when creating PieceAbstract objects. These files do in fact exist,
and the program compiles and runs just fine.

This commit is solely to work around the fact that when running JUnit
tests files for some reason can't be read (always throwing an
IOException).